+++ src/main.rs
@@ -167,7 +167,7 @@ async fn main() -> std::io::Result<()> {
             // .route("/share/{appid}/{shareid}", web::post().to(router::share))
             // .route("/search/{appid}/{tableid}/{viewid}/{id}", web::post().to(router::search))
             // .route("/filter/{appid}/{tableid}/{viewid}", web::post().to(router::filter))
-            // .route("/sort/{appid}/{tableid}/{viewid}", web::post().to(router::sort))
+            .route("/sort/{appid}/{tableid}/{viewid}", web::post().to(router::sort))
             // Gestion du lexique
             .route("/lexic/action/{action}", web::get().to(router::lexicall))
             // .route("/lexic/document", web::get().to(router::lexic))