+++ static/js/silex.js
@@ -172,7 +172,7 @@ $(document).ready(function () {
         var $datas = new FormData();
         var $url = $content.find('.crud-search-input').data("url");
         $datas.append("search", $content.find('.crud-search-input').val().toLowerCase());
-        $datas.append("_xsrf", $("#xsrf").val());
+        // $datas.append("_xsrf", $("#xsrf").val());
         var request =
             $.ajax({
                 type: "POST",
@@ -210,6 +210,254 @@ $(document).ready(function () {
         }
     });
 
-    // Dropdown menu
+    // TRI COLONNE DE LA TABLE
+    $(document).on('click', '.crud-ajax-sort', function (event) {
+        var $sortdirection = "ascending"
+        if (!$(this).hasClass('sorted')) {
+            $(this).closest('tr').find('.sorted').removeClass('sorted');
+            $(this).closest('tr').find('.ascending').removeClass('ascending');
+            $(this).closest('tr').find('.descending').removeClass('descending');
+            $(this).addClass("sorted")
+            $(this).addClass($sortdirection)
+        } else {
+            // on inverse le tri
+            if ($(this).hasClass('ascending')) {
+                $(this).closest('tr').find('.ascending').removeClass('ascending');
+                $sortdirection = "descending"
+                $(this).addClass($sortdirection)
+            }
+        }
+        var $sortid = this.id.substring(4)
+
+        var $datas = {"sortid": $sortid, "sortdirection": $sortdirection};
+        var $node = $(this).closest('table');
+        // $datas.append("sortid", $sortid);
+        // $datas.append("sortdirection", $sortdirection);
+        // $datas.append("_xsrf", $("#xsrf").val());
+        var request =
+            $.ajax({
+                type: "POST",
+                url: "/sort/" + $node.data("appid") + "/" + $node.data("tableid") + "/" + $node.data("viewid"),
+                data: JSON.stringify($datas),
+                dataType: 'json',
+                cache: false,
+                // contentType: 'multipart/form-data',
+                // contentType: false,
+                contentType: "application/json; charset=utf-8",
+                processData: false,
+                beforeSend: function (request) {
+                    //Code à jouer avant l'appel ajax en lui même
+                    console.log(request);
+                }
+            });
+        request.done(function (response) {
+            //Code à jouer en cas d'éxécution sans erreur
+            console.log(response);
+        });
+        request.fail(function (response) {
+            //Code à jouer en cas d'éxécution en erreur
+            console.log(response);
+        });
+        request.always(function () {
+            //Code à jouer après done OU fail dans tous les cas
+            window.location.reload();
+        });
+
+        event.preventDefault();
+    });
+    // TRI retour au tri par défaut de la vue CLICK-DROIT
+    $(".crud-unsort").on("press", function (event) {
+        var $datas = new FormData();
+        var $node = $(this).closest('table');
+        $datas.append("sortid", "");
+        $datas.append("sortdirection", "");
+        // $datas.append("_xsrf", $("#xsrf").val());
+        var request =
+            $.ajax({
+                type: "POST",
+                url: "/sort/" + $node.data("appid") + "/" + $node.data("tableid") + "/" + $node.data("viewid"),
+                data: $datas,
+                dataType: 'json',
+                cache: false,
+                contentType: false,
+                processData: false,
+                beforeSend: function () {
+                    //Code à jouer avant l'appel ajax en lui même
+                }
+            });
+        request.done(function (response) {
+            //Code à jouer en cas d'éxécution sans erreur
+            console.log(response);
+        });
+        request.fail(function (response) {
+            //Code à jouer en cas d'éxécution en erreur
+            console.log(response);
+        });
+        request.always(function () {
+            //Code à jouer après done OU fail dans tous les cas
+            window.location.reload();
+        });
+
+        event.preventDefault();
+    });
+
+    // ACTION DEMANDE CONFIRMATION
+    $('.crud-jquery-action').on('click', function (event) {
+        var $url = $(this).data('url');
+        if ($(this).data('confirm') == true) {
+            $('#crud-action').html($(this).html());
+            $('#crud-modal-confirm')
+                .modal({
+                    closable: false,
+                    onDeny: function () {
+                        return true;
+                    },
+                    onApprove: function () {
+                        $('#beeForm').attr('action', $url);
+                        $('#beeForm', document).submit();
+                    }
+                }).modal('show');
+        } else {
+            // Sans demande de confirmation
+            $('#beeForm').attr('action', $url);
+            $('#beeForm', document).submit()
+        }
+        event.preventDefault();
+    });
+
+    // query sql en ajax
+    $('.crud-jquery-ajax').on('click', function (event) {
+        var $datas = new FormData();
+        var $url = $(this).data('url');
+        // var xsrf = $("#xsrf").val();
+        // $datas.append("_xsrf", xsrf);
+        $.ajax({
+            type: "POST",
+            url: $url,
+            data: $datas,
+            dataType: 'json',
+            cache: false,
+            contentType: false,
+            processData: false,
+        })
+            .done(function (data) {
+                // code en cas de succès
+            })
+            .fail(function (error) {
+                // code en cas d'échec
+            })
+            .always(function () {
+                // code systématique
+                window.location.reload(true);
+            });
+        event.preventDefault();
+    });
+
+
+    // Exécute en ajax une requête SQL sur le serveur et remplit les champs reçus du formulaire courant
+    $('.crud-ajax-sql').on('click', function (event) {
+        var $datas = new FormData();
+        var $url = $(this).data('url');
+        // ajout de données variables à la requête POST
+        var $dataset = $(this).data();
+        for (var d in $dataset) {
+            if (d == "url") continue;
+            var val = $("#" + $dataset[d]).val()
+            $datas.append(d, val);
+        }
+        // var xsrf = $("#xsrf").val();
+        // $datas.append("_xsrf", xsrf);
+        $.ajax({
+            type: "POST",
+            url: $url,
+            data: $datas,
+            dataType: 'json',
+            cache: false,
+            contentType: false,
+            processData: false,
+        })
+            //Ce code sera exécuté en cas de succès - La réponse du serveur est passée à done()
+            //On peut par exemple convertir cette réponse en chaine JSON et insérer cette chaine dans un div id="res"
+            .done(function (data) {
+                if (data.Response != "ok") {
+                    $.toast({
+                        message: data.Message,
+                        class: 'error',
+                        className: {
+                            toast: 'ui message'
+                        },
+                        position: 'bottom center',
+                        minDisplayTime: 1500
+                    });
+                } else {
+                    // mise à jour des rubriques trouvées dans la table
+                    for (var rub in data.Dataset) {
+                        if ($('#' + rub).is("select")) {
+                            $('#' + rub).dropdown('set selected', data.Dataset[rub]);
+                        } else {
+                            $('#' + rub).val(data.Dataset[rub])
+                        }
+
+                    }
+                    $.toast({
+                        message: data.Message,
+                        class: 'success',
+                        className: {
+                            toast: 'ui message'
+                        },
+                        position: 'bottom center',
+                        minDisplayTime: 1500
+                    });
+                }
+                //$("div#res").append(mes);
+            })
+            //Ce code sera exécuté en cas d'échec - L'erreur est passée à fail()
+            //On peut afficher les informations relatives à la requête et à l'erreur
+            .fail(function (error) {
+                let mes = JSON.stringify(error);
+                alert(mes)
+                $.toast({
+                    message: "La requête s'est terminée en échec. Infos : " + JSON.stringify(error),
+                    class: 'error',
+                    className: {
+                        toast: 'ui message'
+                    },
+                    position: 'bottom center',
+                    minDisplayTime: 1500
+                });
+            })
+        //Ce code sera exécuté que la requête soit un succès ou un échec
+        // .always(function () {
+        //   setTimeout(() => { window.location.reload(true) }, 1500);
+        // });
+        event.preventDefault();
+    });
+
+    // UI SEMANTIC
     $('.ui.dropdown').dropdown();
+    $('.message .close')
+        .on('click', function () {
+            $(this)
+                .closest('.message')
+                .transition('fade')
+                ;
+        }
+        );
+    $('.hide')
+        .on('click', function () {
+            $(this)
+                .closest('.message')
+                .transition('fade')
+                ;
+        }
+        );
+
+    // Toaster
+    $('#toaster')
+        .toast({
+            class: $('#toaster').data('color'),
+            position: 'bottom center',
+            message: $('#toaster').val()
+        });
+
 });