+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - bouton coller le presse-papier
 - arg nécessaire dans l'url si view dans un formulaire pour mémoriser le retour
 
+0.16.0 - .. septembre 2023
+- `added` onglets view dans le footer
+- `added` style class sqlite sur les tr
+- `added` tri des colonnes
+
 0.15.0 - 16 septembre 2023
 - `added` portail & view stylisés